Split off SqlCSVScope (MLDB-2030)
