+++ plugins/importtext_procedure.cc
@@ -25,6 +25,7 @@
 #include "mldb/utils/log.h"
 #include "mldb/ext/re2/re2/re2.h"
 #include "mldb/utils/possibly_dynamic_buffer.h"
+#include "sql_csv_scope.h"
 
 
 using namespace std;
@@ -141,215 +142,6 @@ ImportTextConfigDescription::ImportTextConfigDescription()
         }
     };
 }
-
-
-/*****************************************************************************/
-/* SQL CSV SCOPE                                                             */
-/*****************************************************************************/
-
-/** This allows an SQL expression to be bound to a parsed CSV row, which
-    allowing it to find the variables, etc.
-*/
-
-struct SqlCsvScope: public SqlExpressionMldbScope {
-
-    struct RowScope: public SqlRowScope {
-        RowScope(const CellValue * row, Date ts, int64_t lineNumber,
-                 int64_t lineOffset)
-            : row(row), ts(ts), lineNumber(lineNumber), lineOffset(lineOffset)
-        {
-        }
-
-        const CellValue * row;
-        Date ts;
-        int64_t lineNumber;
-        int64_t lineOffset;
-        const RowPath * rowName;
-    };
-
-    SqlCsvScope(MldbServer * server,
-                const std::vector<ColumnPath> & columnNames,
-                Date fileTimestamp, Utf8String dataFileUrl)
-        : SqlExpressionMldbScope(server), columnNames(columnNames),
-          fileTimestamp(fileTimestamp),
-          dataFileUrl(std::move(dataFileUrl))
-    {
-        columnsUsed.resize(columnNames.size(), false);
-        lineNumberUsed = false;
-    }
-
-    /// Column names passed in to the scope
-    const std::vector<ColumnPath> & columnNames;
-
-    /// Which columns are accessed by the bound expressions?
-    std::vector<int> columnsUsed;
-
-    /// Is the line number required by the bound expression?  Some optimizations
-    /// can be turned off if not.
-    bool lineNumberUsed;
-
-    /// What is the timestamp for the actual file itself?  This is used as a
-    /// default timestamp on values returned.
-    Date fileTimestamp;
-
-    /// What is the URI for this file?
-    Utf8String dataFileUrl;
-
-    virtual ColumnGetter doGetColumn(const Utf8String & tableName,
-                                     const ColumnPath & columnName)
-    {
-        if (!tableName.empty()) {
-            throw HttpReturnException(400, "Unknown table name in import.text procedure",
-                                      "tableName", tableName);
-        }
-
-        int index = std::find(columnNames.begin(), columnNames.end(), columnName)
-            - columnNames.begin();
-        if (index == columnNames.size())
-            throw HttpReturnException(400, "Unknown column name in import.text procedure",
-                                      "columnName", columnName,
-                                      "knownColumnNames", columnNames);
-
-        columnsUsed[index] = true;
-
-        return {[=] (const SqlRowScope & scope,
-                     ExpressionValue & storage,
-                     const VariableFilter & filter) -> const ExpressionValue &
-                {
-                    auto & row = scope.as<RowScope>();
-                    return storage = ExpressionValue(row.row[index],
-                                                     row.ts);
-                },
-                std::make_shared<AtomValueInfo>()};
-    }
-
-    GetAllColumnsOutput
-    doGetAllColumns(const Utf8String & tableName,
-                    const ColumnFilter& keep)
-    {
-        vector<ColumnPath> toKeep;
-        std::vector<KnownColumn> columnsWithInfo;
-        size_t numToKeep = 0;
-
-        for (unsigned i = 0;  i < columnNames.size();  ++i) {
-            const ColumnPath & columnName = columnNames[i];
-            ColumnPath outputName(keep(columnName));
-
-            bool keep = !outputName.empty();
-            toKeep.emplace_back(outputName);
-            if (keep) {
-                columnsUsed[i] = true;
-                columnsWithInfo.emplace_back(outputName,
-                                             std::make_shared<AtomValueInfo>(),
-                                             COLUMN_IS_DENSE);
-                ++numToKeep;
-            }
-        }
-
-        // Fill out the offset so we know where it is in the input
-        for (size_t i = 0;  i < columnsWithInfo.size();  ++i) {
-            columnsWithInfo[i].offset = i;
-        }
-        
-        auto exec = [=] (const SqlRowScope & scope, const VariableFilter & filter)
-            {
-                /* 
-                   The filter parameter here is not used since this context is
-                   only used when importing tabular data and there is no way to 
-                   specify a timestamp for this data.
-                */
-
-                auto & row = scope.as<RowScope>();
-
-                RowValue result;
-                result.reserve(numToKeep);
-
-                for (unsigned i = 0;  i < columnNames.size();  ++i) {
-                    if (!toKeep[i].empty())
-                        result.emplace_back(columnNames[i], row.row[i], row.ts);
-                }
-
-                ExcAssertEqual(result.size(), numToKeep);
-
-                return result;
-            };
-
-        GetAllColumnsOutput result;
-        result.exec = exec;
-        result.info = std::make_shared<RowValueInfo>(std::move(columnsWithInfo),
-                                                     SCHEMA_CLOSED);
-        return result;
-    }
-
-    virtual BoundFunction
-    doGetFunction(const Utf8String & tableName,
-                  const Utf8String & functionName,
-                  const std::vector<BoundSqlExpression> & args,
-                  SqlBindingScope & argScope)
-    {
-        if (functionName == "lineNumber") {
-            lineNumberUsed = true;
-            return {[=] (const std::vector<ExpressionValue> & args,
-                         const SqlRowScope & scope)
-                    {
-                        auto & row = scope.as<RowScope>();
-                        return ExpressionValue(row.lineNumber, fileTimestamp);
-                    },
-                    std::make_shared<IntegerValueInfo>()
-                };
-        }
-        else if (functionName == "rowHash") {
-            lineNumberUsed = true;
-            return {[=] (const std::vector<ExpressionValue> & args,
-                         const SqlRowScope & scope)
-                    {
-                        auto & row = scope.as<RowScope>();
-                        if(!row.rowName) {
-                            throw MLDB::Exception("rowHash() not available in this scope");
-                        }
-                        return ExpressionValue(row.rowName->hash(), fileTimestamp);
-                    },
-                    std::make_shared<IntegerValueInfo>()
-                };
-        }
-        else if (functionName == "fileTimestamp") {
-            return {[=] (const std::vector<ExpressionValue> & args,
-                         const SqlRowScope & scope)
-                    {
-                        return ExpressionValue(fileTimestamp, fileTimestamp);
-                    },
-                    std::make_shared<TimestampValueInfo>()
-                };
-        }
-        else if (functionName == "dataFileUrl") {
-            return {[=] (const std::vector<ExpressionValue> & args,
-                         const SqlRowScope & scope)
-                    {
-                        return ExpressionValue(dataFileUrl, fileTimestamp);
-                    },
-                    std::make_shared<Utf8StringValueInfo>()
-                };
-        }
-        else if (functionName == "lineOffset") {
-            return {[=] (const std::vector<ExpressionValue> & args,
-                         const SqlRowScope & scope)
-                    {
-                        auto & row = scope.as<RowScope>();
-                        return ExpressionValue(row.lineOffset, fileTimestamp);
-                    },
-                    std::make_shared<IntegerValueInfo>()
-                };
-        }
-        return SqlBindingScope::doGetFunction(tableName, functionName, args,
-                                              argScope);
-    }
-
-    static RowScope bindRow(const CellValue * row, Date ts,
-                            int64_t lineNumber, int64_t lineOffset)
-    {
-        return RowScope(row, ts, lineNumber, lineOffset);
-    }
-};
 
 
 /*****************************************************************************/
