+++ plugins/plugins.mk
@@ -57,7 +57,9 @@ LIBMLDB_BUILTIN_PLUGIN_SOURCES:= \
 	mock_procedure.cc \
 	datasetsplit_procedure.cc \
 	behavior_dataset.cc \
-	binary_behavior_dataset.cc
+	binary_behavior_dataset.cc \
+	sql_csv_scope.cc \
+
 
 # Needed so that Python plugin can find its header
 $(eval $(call set_compile_option,python_plugin_loader.cc,-I$(PYTHON_INCLUDE_PATH)))
